Add people search autocomplete to components
